Add System.Data.SqlClient 4.8.5 and remove 4.8.3 (#480)
* System.Data.SqlClient 4.8.5

* Update generated code to match previous content
