+++ src/referencePackages/src/system.data.sqlclient/4.8.5/ref/net461/System.Data.SqlClient.cs
@@ -25,11 +25,11 @@ using System.Security;
 [assembly: AssemblyCompany("Microsoft Corporation")]
 [assembly: AssemblyProduct("Microsoft® .NET Framework")]
 [assembly: AssemblyCopyright("© Microsoft Corporation.  All rights reserved.")]
-[assembly: AssemblyFileVersion("4.700.21.41603")]
-[assembly: AssemblyInformationalVersion("4.700.21.41603 built by: SOURCEBUILD")]
+[assembly: AssemblyFileVersion("4.700.22.51706")]
+[assembly: AssemblyInformationalVersion("4.700.22.51706 built by: SOURCEBUILD")]
 [assembly: CLSCompliant(true)]
 [assembly: AssemblyMetadata("", "")]
-[assembly: AssemblyVersion("4.6.1.3")]
+[assembly: AssemblyVersion("4.6.1.5")]
 
 [assembly: TypeForwardedTo(typeof(System.Data.SqlDbType))]
 [assembly: TypeForwardedTo(typeof(System.Data.SqlTypes.INullable))]
